Fixed numEnemies
Finally fixed for good by having each enemy call a GameManager EnemyDied method, which decides if there are no more enemies. Enemies call that when killed, not when hit, so don't have to account for extra enemy lives. Desperately need to think about inheritance, but I think that's too long term a goal for this project.